Always get the estimate primitive count from the attribute
Previously were checking that the attribute was actually but due to
refactoring the shader is not yet examined at this point.
With this change, we always get the estimated number of points from
the attributes.  Potentially it means we're getting it from an attribute
that is not being used but the user should then set the correct number
on the geometry renderer anyway.

Task-number: QTBUG-80697
Change-Id: Ie34131fe3ff41b34609a2f2eb95c4ff678424036
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>