Silence deprecation warnings and fix Qt for Python and documentation builds
Merely addding
DEFINES -= BUILD_QT3D_MODULE
to an example results in a deprecation warnings flood despite
the example not using any deprecrated API since for example
the flag operators and usages in the headers cause warnings by itself.

Remove the deprecation from smart pointer typedefs (it should
be sufficient to deprecate the pointee class) and disable warnings
for usages of deprecated enumerations.

This change silences the build of the simple-cpp example with
BUILD_QT3D_MODULE removed, more changes might be required to fix
the report.

Task-number: QTBUG-81058
Change-Id: Ie70c14cfef024a25b6b1330fe013ce57c69dc962
Reviewed-by: Mike Krus <mike.krus@kdab.com>