Add flag to jobs to indicate if they need to run or not
Once aspect jobs are inserted in the queue, they will always run, but
some of them don't actually have anything to do at every frame.

Ideally, when building a graph of jobs, you would include only jobs
that are required. However this makes building the graph more
complicated.

This introduces a method on the private job pimple which indicated if
the job actually has something to do or not.

When the scheduler is ready to run a job, it will call that method. If
there's nothing to do, then the job will be skipped and the dependent
jobs (if any) run instead.

Change-Id: Iaa8d3d064f03d8c0bab407879a9acea292a06784
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>