Skip RenderView which don't need to be submitted
This is when a particular set of filters/layers ... result
in a renderview that has no render commands to submit.
Trying to submit such renderview is just a waste of time.

RenderViews that have no commands but do buffer download,
back buffer clearing, setting fences .... are still submitted
even if there are no commands to render.

Change-Id: I4f66334b9b2170249c7c7aae1024ebccc8427750
Reviewed-by: Mike Krus <mike.krus@kdab.com>