Add public API for working with jobs
Existing jobs used private API for doing work after jobs were run or
controlling if running was required or not.

Make this public so that new jobs in custom aspects don't need to
use private API.

Unfortunately, this requires changing the signature of one public
method that mistakenly was using a private class as a parameter.

Change-Id: Ie8315a0f98ec48f84650163b4ac713e7e6845a92
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>