Allow for when a Scene3D item switches screens
When a Scene3D item switches screens then it will need to be
reinitalized so that the supporting contexts, offscreen surfaces are
set to use the same screen and not the original one. This ensures that
the item is still rendered correctly on the new screen.

This includes a manual test using QQuickWidget in separate windows that
enables it going from one screen to the other.

This is a fresh version after the previous version was found to have a
bug shown in the scene3d-loader test which has now been resolved.

Change-Id: I3c711e894018db52ec00a8a5d2e0fb0128743ab1
Done-with: Antti Kokko <antti.kokko@qt.io>
Fixes: QTBUG-79192
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Reviewed-by: Mike Krus <mike.krus@kdab.com>
(cherry picked from commit 4eef300be70509a208527bf164f7746fa1bf07a1)