Uniform preparation: stop copying the ShaderUniforms around
Those are static and expensive to copy (especially once per RenderCommand)
as they contain several QStrings. Instead we sort them by nameIds when
the shader is initialized and use indices to reference them.

This avoids the copies and the sorting allows to be more efficient
at finding the ShaderUniforms based on a nameId.

Change-Id: I50008381f1d336ab182f5ac5a08500791e2871e9
Reviewed-by: Mike Krus <mike.krus@kdab.com>