Reduce CommandBuilding/Update job spawning threshold
Those jobs are usually the most expensive to run. It therefore makes
sense we leverage parallelism as much as possible and not discard
threads we could be using.

In that sense, the minimum amount of elements to process per job is reduced
from 100 to 10 which means that we will spawn threads if they can contain
at least 10 element to process (instead of 100 previously).

This now allows to benefit from more parallelism on smaller scenes (that have
less than 100 entities) while still scaling similarly as previously on
larger scenes.

Change-Id: I91e97fc23185d02c2727d87bb7a4ea4991f5ff14
Reviewed-by: Mike Krus <mike.krus@kdab.com>