Scene3DRender: connect to beforeRenderPassRecording to perform draw
Previously we were connecting to beforeRendering which happens prior to
RHI having created its render pass.
When RHI creates its render pass, it clears the rendering surface. In the
case Scene3D is used in Underlay this would therefore clear the 3D content.

We now connect to beforeRenderPassRecording which happens after beforeRendering
when RHI has created its RenderPass (and therefore cleared the screen) but
prior to drawing/recording commands for the actual QtQuick rendering.
This allows Qt3D to clear (if FrameGraph has that operation) or not and should
work for both Underlay and FBO rendering modes.

That being said, at the moment, apitrace shows that QtQuick still clears before
drawing after beforeRenderPassRecording what emitted. This lead to Scene3D in
Underlay compositing mode to not be visible.

Change-Id: I55bd5ae9f26926f06e9e50793da316918b78b6b7
Reviewed-by: Mike Krus <mike.krus@kdab.com>