Keep only a single EntityRenderCommandDataView per RenderView
We used to have 2 instances we would flip/flop every frame to maintain
an up to date cache of RenderCommands as we would be submitting one set
while preparing the next once.

Since the RenderThread has now been removed, we now only need to keep a
single instance as we can't be preparing and submitting at the same time.

Change-Id: Ib9c258846efb31dc0afcf047f23d2d9522e05441
Reviewed-by: Mike Krus <mike.krus@kdab.com>