Use a single UBO per binding for all commands using the same pipeline
Easier said than done. Main issue is the fact that RHI combines
ResourceLayoutDescription and ResourceBinding in the same object. This
prevents creating a pipeline solely based on the resource description
(which would make sense if you plan on reusing pipelines), instead you need
to provide buffers to go with the description. That being said, a RHI patch
is on its way to allow providing nullptr for the resources when all we care
about is specifying a layout.

In the mean time, since we need the binding description to create a pipeline,
this forces us to have buffers we will later destroy only to create the pipeline
because we don't know yet how many commands will be used.

Later improvements to be made are:
- Share pipelines based on Geometry layout instead of Geometry NodeId
- Find ways to cache link between RenderCommands and Pipelines
- Handle cases where user directly provided a UBO as a parameter value

Change-Id: Iec989885711d850f672268910f03468149a57f32
Reviewed-by: Mike Krus <mike.krus@kdab.com>