QShaderGenerator: generate node headers in the same order as statements
This ensures that header snippets that depend on one another are forced
to be emitted in the correct order, assuming the nodes also depend on
one another.

This has the side effect of dropping header snippets from nodes which
are not connected to any output node or with any unbound input.

Also removed a few lines of unused code.

Pick-to: 5.15
Change-Id: I5e544470aea1d34467f8165fb49e48d38931e0bc
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>