rhi pipelineuboset: handle more commands uniforms that can fit in a single UBO
We create a pool of UBOs based on a relatively low max UBO size (16384).
For each command, we then create bindings that point to the correct
UBO in the pool with the proper local offset based on the command index.

This allows to handle as much commands using the same pipeline as we can.

Change-Id: I06a8e722090e5f70bd473847e2c308a987b0d94c
Reviewed-by: Mike Krus <mike.krus@kdab.com>