rhi: Rework the way we handle RenderTargets
Instead of creating a RenderTarget when we create a pipeline, we now
create RenderTargets any time a frontend defines one.

This allows to align with the other type of resources that we create
before pipelines.

Add a dirty flag on QRenderTarget so that backend can easily check if a
frontend RenderTarget has changed.

Also remove the handling of QRenderTargetSelector::outputs which allows
to specify a subset of draw buffers for a given FBO. This cannot be
handled with RHI and in a more general way, removing that altogether
might simplify the API. If you share the attachments, nothing stops
you from creating several FBO to handle that case.

Fix coding style and update the unit tests.

Change-Id: I9f3623e9d7ab9d508db207f2c43cc6c1eedb4f01
Reviewed-by: Mike Krus <mike.krus@kdab.com>