RenderCommands: try to resolve API shader at command build time
This ensures that if a commands keeps being rebuilt, it will be usage
for the current frame, assuming the shader exists.

This aligns the RHI renderer with what the GL renderer does.
The instanced-arrays-qml is a good way to verify that a command
that keeps being rebuilt is rendered correctly.

Change-Id: Ie9e574e98efeec6eb836af613d28cbf4cdf2fef5
Reviewed-by: Mike Krus <mike.krus@kdab.com>