Update frontend data after job
Data computed by the core aspect needs to be available for the
render aspect when it runs. This forces the update to the front node
from the job thread which should be save currently since the main thread
is blocked, there's no other core aspect jobs running, and the render
aspect job depends on this one.

Change-Id: I9483066d272eec02456f9c02f8e4a34acad35e0b
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>