Ensure Scene3D re-renders when a backend node is updated
There are two ways in which backend nodes can receive changes from the
frontend:

1) By syncing the state directly from the frontend nodes
2) By changs delivered by the change arbiter

Scene3D is already equipped to deal with changes going via the first
mechanism. However, it is possible for changes going via the second
mechanism to not trigger a re-render of a Scene3D item using the
OnDemand render mode. This is even true in the case of using a Scene3D
inside of a QQuickWidget even though this disables the threaded
Qt Quick 2 rendering.

The way this manifests is that:

* A change is made via mechanism 2) such as creating some new nodes.
* The frontend changes trigger a re-render of the Scene3D item via the
  existing notification mechanism (connection to the change arbiter's
  receivedChange() signal.
* The QQuickWidget uses a timerEvent() to trigger a possible update
* This notices the Scene3D is dirty and causes it to render and mark
  its state as clean. This does not yet pick up the newly created
  frontend nodes though as they need to go through the deferred
  post creation init dance.
* The change arbiter then processes the deferred node creation events
  and creates the backend nodes.
* At the next timerEvent from the QQuickWidget it still sees that the
  Scene3D is clean and does not bother to re-render it.

This change fixes this by adding a new signal to the change arbiter,
syncedChanges() that gets emitted only when the arbiter has
actually finished distributing the changes going via mechanism 2) to
the backend nodes.

This syncedChhanges() signal is emitted on the aspect thread and so
Scene3DItem connects to this using a queued connection.

This then ensures that once the new backend nodes are actually
present and up to date that the Scene3D in OnDemand mode will then
re-render the scene with the complete scene graph.

Change-Id: I3246f34a3cfe15392a390eb6d97b4e90346288a6
Reviewed-by: Mike Krus <mike.krus@kdab.com>