Scene3DItem: release frontend/backend tree on destruction
We fully release the AspectEngine when the last of the threads between
main/sg thread is destroyed. We do it that way as the destruction order
of the threads is non deterministic and we can only destroy the
AspectEngine after it has been shut down from the SGThread.

However, that can lead to the Entity tree to never be properly
released/cleaned up in the case where the main thread is destroyed before
the render thread.

To account for that, we now release the root entity when the
AspectEngineDestroyer allowRelease is called in the context of the main
thread. That way, even if the AspectEngine is fully destroyed later
once SG thread exits(), the scene tree is still cleaned up while we have
an event loop.

Special handling is done for that behavior not to happen in the case
Scene3D is rendered from with a QQuickRenderControl.

This should fix asserts or crash observed when using Scene2D within a
Scene3D scene.

Change-Id: I77edff567d60e53a61023f88b551fc475885e5eb
Reviewed-by: Mike Krus <mike.krus@kdab.com>