QShaderProgram: handle binding = auto replacement for raw shader code
binding = auto replacement only took place when using shader graphs.
This meant providing shader code that used binding = auto or that included
snippets that themselves used binding = auto (such as phong.inc.frag)
would fail.

For that reason, that replacement was moved from the shadergraph builder
to QShaderProgramPrivate so that it can be used for both cases: raw
shader code or shader graphs.

Change-Id: I298a82b83b1181d7a0f19a2e2fefc3aeb9eb81c4
Pick-to: 6.0 6.0.0
Reviewed-by: Mike Krus <mike.krus@kdab.com>