Scene3D: handle flipping based on the RHI backend in use
Qt3D handles flipping based on whether the RHI backend in use requires
it or not. Scene3D would also flip the content (ignoring the fact that
the content was already in the correct orientation).

This resulted in cases where a Qt 3D scene rendered with Scene3D would be
flipped while the same scene rendered on its own (with Qt3DWindow) would
be correct.

Change-Id: I9fc2a01f63713ed3dbafaf896dfd9457d7160653
Task-number: QTBUG-87842
Pick-to: 6.0 6.0.0
Reviewed-by: Mike Krus <mike.krus@kdab.com>