OpenGL renderer: fix use of outdated VAO when shader is reloaded
In a case where an existing shader was updated, VAO previously associated
with the shader wouldn't be updated properly.

To account for this, we now release all VAOs associated with a shader when
loading a shader. Additionally we also skip VAO creation of commands referencing
a shader that was loaded in the current frame. The reason for that is we know
the command for that VAO is not completed. Qt3D requires a first frame to load a
shader + perform introspection and creates valid command against shader only on
the following frame.

Change-Id: I36f89fdd78e857dc5bc4af3b539d3b32630dfad6
Reviewed-by: Mike Krus <mike.krus@kdab.com>
(cherry picked from commit 7f1bcdb799ad689fc138f976eed7e12f1ce64386)