Fix binding of CubeMapArray textures
To bind to a specific face: use glFramebufferTextureLayer and calculate
the layer as 6 * layer + face.

To bind to a specific cubemap: There is no way in OpenGL to directly
bind  to an entire cubemap within the array.  Instead bind to the entire
texture as for a CubeMap with AllFaces. It is up to the drawing code
to select the right cubemap and face, 6 * layer + face

Change-Id: Ib9e1a598503eeba1fa729677944aba6d799f90b2
Pick-to: 5.15 6.0
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>