Fix feature names in cmake
Feature names were mistakenly changed to use underscores rather
than dashes. This breaks builds of apps using qmake as it changes
the feature detection symbols.

Fixed by reverting to use dashes in the names. This is fine in cmake
as dashes are converted to underscores when not quoted.

Unfortunately, it changes the command line configuration arguments back
to the dashed versions.

[ChangeLog] Configuration time command line arguments have changed to
use dashes rather than underscores

Change-Id: If793c5424aa0459daf06dd075e2e40c33997ca1a
Task-number: QTBUG-89386
Pick-to: 6.0
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>