Make Qt3DWindow handle default surface format more gracefully
The default is now to use the same rules as QtQuick.
If nothing is specified, will use RHI and the best surface for the
platform (Metal on Mac, DX on Windows, Vulkan if available, etc).

If Qt3D's OpenGL backend is requested via QT3D_RENDERER, will always
use OpenGL surface.

If QSG_RHI_BACKEND is set, will honour that if possible.

Otherwise will use the API requested in the constructor (which
defaults to using auto detection).

Pick-to: 6.0
Change-Id: I8941a081940c073dfcf66711a8628e8ef6fd0915
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>