conanfile.py: Use QTDIR instead of QT_PATH
There's an established environment variable for this though, QTDIR.
This environment variable is also set by Qt Creator by default.
So the Conan recipes should accept this environment variable instead.

This requires a change in the matching conan profile files
where the path to Qt installation is injected to env.

Task-number: QTBUG-90583
Change-Id: I0beaff832325008dd1778a03d2885b6190def754
Reviewed-by: Kai Koehne <kai.koehne@qt.io>