Conan: Use the 'shared' option value of the Qt build
Default to the value of the prebuilt Qt build.

The 'default' needs to be added as valid option in the list which
will be used by default if no value is given by the user or via
the conan profile.

The user is still able to override this via the profile or from
command line if needed:

  "conan install ... -o shared=False"

CMakeCache.txt:

  - "conan install ..."
    - BUILD_SHARED_LIBS:BOOL=ON
  - "conan install ... -o shared=True"
    - BUILD_SHARED_LIBS:BOOL=ON
  - "conan install ... -o shared=False"
    - BUILD_SHARED_LIBS:BOOL=OFF

Pick-to: 6.1
Task-number: QTBUG-90901
Change-Id: Ifd9ef0d982a1f1e4a3f454e3d5e4c7075f873c22
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>