Fix doc conflict about FilterKey of RenderPass and RenderPassFilter
RenderPass doc previously said that the RenderPass is executed "when no
FilterKey nodes are specified and no RenderPassFilter is present in the
FrameGraph." It can be read that RenderPass having no FilterKey nodes
should be rejected if RenderPassFilter is present.

  foundMatch = !passFilter && renderPass->filterKeys().isEmpty();

On the other hand, RenderPassFilter doc says "if no RenderPassFilter
is present, then all RenderPass objects are selected for rendering."

  foundMatch = !passFilter;

Since the actual behavior is the latter, this patch adjusts the
RenderPass doc.

If I read it correctly, the condition in findRenderPassesForTechnique()
is as follows, so the current implementation (i.e. true if empty) makes
sense.

  all(any(p == f for p in renderPass.filterKeys)
      for f in passFilter.filters)
  # or renderPass.filterKeys is superset of passFilter.filters

Change-Id: I41d53ce4cab022e93f99cc20a1a431e97b9fab49
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>