Fix detection of Vulkan
When Qt is built to enable Vulkan support, it is given the path to
the Vulkan header source. This then enables the vulkan feature.

Qt3D used this feature to decide if it could use Vulkan directly or
not, however, since it is now built separately, it has no idea where
to find the headers, so QVulkanInstance is not properly defined.
This caused an issue when using official installers of Qt since
those are configured to support Vulkan.

So we now do a separate config check to find Vulkan for Qt3D.

When building Qt3D, you need to make sure if can find Vulkan (the same
way you configure Qt itself)

Task-number: QTBUG-90243
Change-Id: Ia05381597b4974d7b0914706ea2f607341dccacc
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
(cherry picked from commit 9a473a3c9b246f4895ae73d7060745b8b199a6c5)