Transform Qt3D.Core to declarative type registration
We already did this for the value types. As we cannot mix declarative
and procedural registration, we need to do it for the object types, too.

For declarative registration, the qmltypes files are generated at
compile time. For procedural registration, you need to manually invoke
qmlplugindump to generate them. Doing both leads to two conflicting
qmltypes files for the same module.

Fixes: QTBUG-92259
Pick-to: 6.1
Change-Id: Ic3d9c63410f84e04cd368ddaee3a6ff2d9a3c997
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mike Krus <mike.krus@kdab.com>