Fix build on macOS if system assimp is installed
On macOS, if pkg-config is not used, we modify
CMAKE_SYSTEM_FRAMEWORK_PATH and CMAKE_SYSTEM_PREFIX_PATH.  This must be
done in the configure tests too.  Otherwise the find_package commands
may return different things for configure tests and the Qt build itself.

In this case, assimp was picked up from brew in the configure test but
not in the Qt build.

Task-number: QTBUG-97076
Change-Id: Ib43ceebe78768267a60142c5a4235eab7ac18c81
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>