RHI: Fix content not updated when using OnDemand and switching textures
When changing a Parameter whose value is a texture, it takes Qt3D potentially
two frames to process the scene fully before it is able to draw the updated
content.

This patch makes it so that if we know we have the flags ParametersDirty for
frame N, we request to reprocess the scene once more on frame N+1.

Change-Id: I9df868ce2aac857bad1a1cb1074038b9c5bcd907
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
(cherry picked from commit 9bdabc8c1412ae58080f700f42f2393c6357788a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>