Update commercial license headers
Updated header.COMM to the files in tqtc-qt3d.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4935
Change-Id: I769e2e6a9afe8a340fb04ae002cd37a3b356c701
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>