Qt3D: replace qSwap with member-swap where possible
qSwap() is a monster that looks for ADL overloads of swap() and also
detects the noexcept of the wrapped swap() function, so it should only
be used when the argument type is unknown. In the vast majority of
cases, the type is known to be efficiently std::swap()able or to have
a member-swap. Call either of these.

Task-number: QTBUG-97601
Pick-to: 6.3 6.2
Change-Id: I1f27dc458f2d57a71070e818583f2e90c82e600d
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>