Fix Light Uniform handling
* OpenGL Renderer:
When we parse shaders we record standard uniform names (e.g mvp),
light uniform names and other unknown uniform names in dedicated lists.
However, when setting the light uniform values, instead of using the light uniform names lists
to check whether a value could be set or not, we were using the unknown uniform names list and
therefore, we ended up not setting the value.

* RHI Renderer:
When parsing the shaders to record uniform names, even though we iterate through uniform struct
arrays and members, we were not recording the member names as uniform names. This prevented
from ever being able to set a the value of a struct or array member uniform.
Additionally, the way we handled introspecting and uploading UBOs made up of arrays of structs
was incorrect.

Task-number: QTBUG-95650
Task-number: QTBUG-100402
Pick-to: 6.4 6.2 6.0
Change-Id: I7e33575906c1bb376fe8059174c09cda553930e4
Reviewed-by: Mike Krus <mike.krus@kdab.com>