RHI backend: set texture handle on the frontend node
Allows to then use the handle with a 3rd party external renderer
(e.g QtQuick with QQuickWindowPrivate::createTextureFromNativeTexture).

Note: we need to be careful and only set the handles once we know the internal
QRhiTextures aren't modified anymore. This means we can only do it after
having created/updated the textures and the render targets as the render
target creation might result in some QRhiTextures being recreated.

Pick-to: 6.4
Change-Id: I7e85fd2661cf935368ee55dc2ebd240bbda38a6f
Reviewed-by: Mike Krus <mike.krus@kdab.com>