Scene3DRenderer: reset OpenGL state when skipping frames
When using OnDemand rendering, Qt3D/Scene3DRenderer will only
render if anything in the scene has changed.

The OpenGL state is reset before Qt3D renders and after it has rendered to leave
a clean GL state for QtQuick to render.

However, when skipping frames, we never enter the render function and therefore
never reset the OpenGL state that QtQuick will be using.

Therefore, it could happen when skipping Qt3D frames that the last time Qt3D
rendered or prepared a frame, it left some objects bound in the GL state.
Since the GL state isn't reset in that case, it breaks assumptions made by
QtQuick (e.g it might expect texture unit 0 to contain a glyph texture and not
whatever Qt3D might have left bound to it). This could result in incorrect
rendering of the QtQuick content.

This patch forces a GL state reset when Qt3D is asked to skip frames. It will
only reset the GL state for the first Qt3D frame being skipped as there's no
point in doing it for each consecutive skipped frame.

Task-number: QTBUG-105376
Change-Id: I281c2196a60f5aa8f316f4a2d0b9ddebec275188
Reviewed-by: Mike Krus <mike.krus@kdab.com>