Unpin previously used image units during render command parameter setup
Image units are not unpinned after executing a render/compute command and before executing the next one.
This can lead to failure during binding of textures to image units if the number of required image units of subsequent commands exceeds the overall number of image units.
Therefore, the image units are now unpinned during parameter setup of each command.

Pick-to: 6.4 5.15
Change-Id: I9ae7be0bf8c2efe33aca836ef9e81db7c8f5b654
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>