ShaderParameterPack: check we only store a single UBO/SSBO per BlockBinding
Since we cache ShaderParameterPacks over multiple frames, we need to
ensure that we check that when setting a UBO/SSBO we don't already
have an entry matching the block index.
If that's the case we should simply replace the entry rather than
append a new entry. Failing to do so results in appending a growing
number of UBO entries for the same block.

Change-Id: Ie7535a80b9426ce2fd5ab6ebe16ea71bd7447750
Reviewed-by: Mike Krus <mike.krus@kdab.com>
(cherry picked from commit f672753556cf28d44f864457dc6338a55e2e0ed4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>