OpenGL SubmissionContext: reset defaultFBO every time we begin drawing
We wrongly assumed that the default FBO id would remain constant over time
once the submission context had been initialized. However it appears that
when using Qt3D with a QQuickWidget/QQuickRender control with offscreen
surfaces, the default FBO might change over time as the offscreen surface
might get recreated.

To compute the proper render target size, we rely on comparing the default FBO
id against the active FBO id being used at various times in the rendering
process.

If the default FBO has changed since initialization (due to an offscreen windows
recreation on a resize for example), we run the risk of having invalid
comparisons between the default FBO and current active FBO.
This results in possible erronous render size values when going through
SubmissionContext::renderTargetSize. This can lead to occasional flickering or
the scene not rendering (getting a reported size of 1x1) when resizing
a QQuickWidget holding a Scene3D.

Pick-to: 6.2 6.4 5.15
Change-Id: I18371e8e19213c472bdb64a6f339ee0941e5cb3d
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>